update master brunch

+++ app/Http/Controllers/Buyer/OrderController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Buyer;
 
 use App\Events\NewOrderStoreEvent;
+use App\Events\OnCancelOrderItem;
 use App\Helpers\OrderHelper;
 use App\Helpers\TemplateHelper;
 use App\Models\BillingInfo;
@@ -49,23 +50,30 @@ class OrderController extends Controller
         }
     }
 
-    public function show(Request $request, $order_no){
-        $orderInfo = Order::where('order_no', $order_no)->with('buyer.user', 'orderItems.product','orderItems.seller', 'billing', 'shipping', 'payment')->first();
-
-        if($request->ajax()){
-            if(!empty($orderInfo)) {
-                return ResponserTrait::singleResponse($orderInfo, 'success', Response::HTTP_OK, 'Order Details Found');
-            }else{
-                return ResponserTrait::allResponse('error', Response::HTTP_NOT_FOUND, 'Order Details Not Found');
-            }
-        }else{
-            if(empty($orderInfo)){
-                return abort(Response::HTTP_NOT_FOUND);
-            }
-            return view('template.'.$this->template_name.'.order.show',[
-                'order_no'=>$order_no,
-            ]);
+    public function show($order_id){
+        $orderInfo = Order::where('order_id', $order_id)->first();
+        if(empty($orderInfo)){
+            return abort(Response::HTTP_NOT_FOUND);
         }
+        return view('templates.'.$this->template_name.'.buyer.order.order_details',[
+            'orderId'=>$orderInfo->order_id,
+        ]);
+
+    }
+
+    public function Order_details($orderId)
+    {
+        $orderInfo = Order::where('order_id', $orderId)
+            ->with(['buyer.user','billing', 'shipping', 'payment', 'orderItems'=>function($query){
+                return $query->with('product', 'seller.shop', 'image');
+            } ])->first();
+
+        if(!empty($orderInfo)) {
+            return ResponserTrait::singleResponse($orderInfo, 'success', Response::HTTP_OK, 'Order Details Found');
+        }else{
+            return ResponserTrait::allResponse('error', Response::HTTP_NOT_FOUND, 'Order Details Not Found');
+        }
+
     }
 
     public function order_store(Request $request){
@@ -75,9 +83,9 @@ class OrderController extends Controller
         }
         // TODO check Validation
         $validator = Validator::make($request->all(),[
-            'billing_address'=>'required|array',
+            'billing_address'=>'array',
             'shipping_address'=>'required|array',
-//            'shipping_method_id'=>'required',
+            'shipping_method_id'=>'required',
             'payment_method_id'=>'required',
         ]);
 
@@ -95,12 +103,12 @@ class OrderController extends Controller
                     'order_no'=>uniqid(),
                     'buyer_id'=>$buyer->buyer_id,
                     'total_qty'=>Cart::count(),
-                    'sub_total'=>Cart::subtotal(),
+                    'sub_total'=>0,
                     'discount'=>$request->discount,
                     'voucher_id'=>$request->voucher_id,
                     'voucher_price'=>$request->voucher_price,
                     'delivery_charge'=>$request->delivery_charge,
-                    'total'=>(Cart::total()+$request->delivery_charge),
+                    'total'=>0,
                     'order_date'=>now(),
                     'order_status'=>config('app.active'),
                     'shipping_method'=>(!empty($request->shipping_method_id))?$request->shipping_method_id:1,
@@ -109,11 +117,12 @@ class OrderController extends Controller
                 if(!empty($order)){
                     $orderItemsArray = [];
                     $productUpdateArray = array();
+                    $orderTotal = 0;
                     // get single cart product
                         foreach ($carts as $cart){
                             // get product, brand, size, color details
                             $product = Product::where('product_id',$cart->id)->with('brand', 'singleVariation')->first();
-                            if($product->product_type == Product::ProductType['Variation']&& !empty($cart->options->sizeId) && !empty($cart->options->colorId)){
+                            if($product->product_type == Product::ProductType['Variation'] && !empty($cart->options->sizeId) && !empty($cart->options->colorId)){
                                 $variationProduct = ProductVariation::where('product_id', $cart->id)
                                     ->where('pri_id', $cart->options->colorId)
                                     ->where('sec_id', $cart->options->colorId)->first();
@@ -136,25 +145,26 @@ class OrderController extends Controller
                                 'variation_id'=>$variation_id,
                                 'status'=>$product->product_status,
                             ]);
-
                             array_push($orderItemsArray,[
                                 'order_id'=>$order->order_id,
                                 'buyer_id'=>$buyer->buyer_id,
                                 'seller_id'=>$product->seller_id,
                                 'product_id'=>$product->product_id,
                                 'product_name'=>$product->product_name,
-                                'size_id'=>(!empty($cart->options->sizeId))?$cart->options->sizeId:null,
-                                'size'=>(!empty($cart->options->size))?$cart->options->size:null,
-                                'color_id'=>(!empty($cart->options->colorId))?$cart->options->colorId:null,
-                                'color'=>(!empty($cart->options->color))?$cart->options->color:null,
+                                'size_id'=>$cart->options->sizeId??null,
+                                'size'=>$cart->options->size??null,
+                                'color_id'=>$cart->options->colorId??null,
+                                'color'=>$cart->options->color??null,
                                 'brand_id'=>$product->brand_id,
                                 'brand'=>(!empty($product->brand))?$product->brand->brand_name:'',
                                 'price'=>$cart->price,
                                 'qty'=>$cart->qty,
                                 'subtotal'=>($cart->qty*$cart->price),
                                 'discount'=>0,
-                                'total_price'=>($cart->qty*$cart->price)
+                                'total_price'=>($cart->qty*$cart->price),
+                                'image_id'=> $cart->options->image_id
                             ]);
+                            $orderTotal += ($cart->qty*$cart->price);
                         }
 
                     // Store order items
@@ -165,27 +175,29 @@ class OrderController extends Controller
                     }
 
                     if(!empty($orderItems)){
+
                         // TODO store order billing details
-                        $billingInfo = (Object) $request->get('billing_address');
+                         $billingInfo = (Object) $request->get('billing_address');
+                        if(!empty($request->get('billing_address'))){
+                            $billing = BillingInfo::create([
+                                'order_id'=>$order->order_id,
+                                'buyer_id'=>$buyer->buyer_id,
+                                'address_id'=>$request->billing_address_id,
+                                'first_name'=>$billingInfo->first_name,
+                                'last_name'=>$billingInfo->last_name,
+                                'phone_no'=>$billingInfo->phone_no,
+                                'address'=>$billingInfo->address,
+                                'city'=>$billingInfo->city,
+                                'state'=>$billingInfo->state,
+                                'district'=>$billingInfo->district,
+                                'region'=>$billingInfo->region,
+                                'postal_code'=>$billingInfo->postal_code,
+                                'country'=>'somalia',
+                            ]);
 
-                        $billing = BillingInfo::create([
-                            'order_id'=>$order->order_id,
-                            'buyer_id'=>$buyer->buyer_id,
-                            'address_id'=>$request->billing_address_id,
-                            'first_name'=>$billingInfo->first_name,
-                            'last_name'=>$billingInfo->last_name,
-                            'phone_no'=>$billingInfo->phone_no,
-                            'address'=>$billingInfo->address,
-                            'city'=>$billingInfo->city,
-                            'state'=>$billingInfo->state,
-                            'district'=>$billingInfo->district,
-                            'region'=>$billingInfo->region,
-                            'postal_code'=>$billingInfo->postal_code,
-                            'country'=>'somalia',
-                        ]);
-
-                        if(empty($billing)){
-                            throw new \Exception('Order Not Place', Response::HTTP_BAD_REQUEST);
+                            if(empty($billing)){
+                                throw new \Exception('Order Not Place', Response::HTTP_BAD_REQUEST);
+                            }
                         }
 
                         // TODO store order shipping details
@@ -256,11 +268,19 @@ class OrderController extends Controller
                         }
                         $user = User::where('user_id', auth()->id())->with('buyer')->first();
                         event(new NewOrderStoreEvent($user,$order));
+                        $order = $order->update([
+                            'sub_total'=>$orderTotal,
+                            'total'=>$orderTotal+$request->delivery_charge
+                        ]);
+                        if(!empty($order)){
+                            DB::commit();
+                            Cart::destroy();
+                            //TODO Change to Invoice Page url
+                            return ResponserTrait::allResponse('success', Response::HTTP_CREATED, 'Your Order Place Successfully', '', route('buyer.order.index'));
+                        }else{
+                            throw new \Exception('Order Not Place. Try Again.', Response::HTTP_BAD_REQUEST);
+                        }
 
-                        DB::commit();
-                        Cart::destroy();
-                        //TODO Change to Invoice Page url
-                        return ResponserTrait::allResponse('success', Response::HTTP_CREATED, 'Your Order Place Successfully', '', route('buyer.order.index'));
                     }else{
                         throw new \Exception('Order Item Not Insert.', Response::HTTP_BAD_REQUEST);
                     }
@@ -276,6 +296,98 @@ class OrderController extends Controller
         }else{
             $errors = array_values($validator->errors()->getMessages());
             return ResponserTrait::validationResponse('validation', Response::HTTP_BAD_REQUEST, $errors);
+        }
+    }
+
+    public function order_item_cancel($item_id)
+    {
+        try {
+            DB::beginTransaction();
+
+            $orderItem = OrderItem::where('item_id',$item_id)->first();
+            $orderId = $orderItem->order_id;
+            if(empty($orderItem)){
+                throw new \Exception('Invalid Order Item Information', Response::HTTP_NOT_FOUND);
+            }
+            if(!empty($orderItem) && $orderItem->item_status !== OrderItem::ItemStatus['Active']){
+                throw new \Exception('Not able To Cancel. Contract With Admin.', Response::HTTP_NOT_ACCEPTABLE);
+            }
+
+            $orderItem = $orderItem->update([
+                'item_status'=>OrderItem::ItemStatus['Cancel'],
+                'cancel_by'=>OrderItem::CancelBy['Buyer'],
+            ]);
+
+            if(!empty($orderItem)){
+                ## TODO Order History Added,
+                ## TODO Check if All Item Are Cancel then update the order status as Cancel
+                event(new OnCancelOrderItem($item_id));
+                $notCancelItem = OrderItem::where('order_id', $orderId)->where('item_status', '!=', OrderItem::ItemStatus['Cancel'])->count();
+                if($notCancelItem == 0){
+                    $order = Order::where('order_id', $orderId)
+                        ->update([
+                            'order_status'=>Order::OrderStatus['Cancel'],
+                        ]);
+                    if(empty($order)){
+                        throw new \Exception('Invalid Information', Response::HTTP_BAD_REQUEST);
+                    }
+                }
+                DB::commit();
+                return ResponserTrait::allResponse('success', Response::HTTP_OK, 'Order Item Cancel Successfully');
+            }else{
+                throw new \Exception('Some Error Found. Try Again.', Response::HTTP_BAD_REQUEST);
+            }
+        }catch (\Exception $ex){
+            DB::rollBack();
+            return ResponserTrait::allResponse('error', $ex->getCode(), $ex->getMessage());
+        }
+    }
+
+    public function order_cancel($orderId)
+    {
+        try {
+            DB::beginTransaction();
+
+            $order = Order::where('order_id', $orderId)->with('orderItems')->first();
+
+            if(empty($order)){
+                throw new \Exception('Invalid Order Information', Response::HTTP_NOT_FOUND);
+            }
+            if(!empty($order) && $order->order_status !== Order::OrderStatus['Active']){
+                throw new \Exception('Not able To Cancel This. Contract With Admin.', Response::HTTP_NOT_ACCEPTABLE);
+            }
+            $totalOrderItem = $order->orderItems->count();
+            $itemStatus = $order->orderItems->where('item_status', OrderItem::ItemStatus['Active'])->count();
+
+            if($totalOrderItem !== $itemStatus){
+                throw new \Exception('Not Able To Cancel this Order. Contract With Admin.', Response::HTTP_NOT_ACCEPTABLE);
+            }
+            $order = $order->update([
+                'order_status'=>Order::OrderStatus['Cancel'],
+            ]);
+
+            if(!empty($order)){
+                $activeItems = OrderItem::where('order_id', $orderId)->where('item_status', OrderItem::ItemStatus['Active'])->get();
+                foreach ($activeItems as $item){
+                    event(new OnCancelOrderItem($item->item_id)); // Update Product qty if cancel order
+                }
+                $items = $activeItems->update([
+                    'item_status'=>OrderItem::ItemStatus['Cancel'],
+                    'cancel_by'=>OrderItem::CancelBy['Buyer'],
+                ]);
+                if(!empty($items)){
+                    ## TODO ADd Order History
+                    DB::commit();
+                    return ResponserTrait::allResponse('success', Response::HTTP_OK, 'Order Item Cancel Successfully');
+                }else{
+                    throw new \Exception('Some Error Found. Try Again.', Response::HTTP_BAD_REQUEST);
+                }
+            }else{
+                throw new \Exception('Some Error Found. Try Again.', Response::HTTP_BAD_REQUEST);
+            }
+        }catch (\Exception $ex){
+            DB::rollBack();
+            return ResponserTrait::allResponse('error', $ex->getCode(), $ex->getMessage());
         }
     }
 }
